+++ src/channel/index.ts
@@ -204,7 +204,7 @@ class Channel implements Channel {
             }
           }
 
-          const error: E.ErrorMessage | void = await tutorialConfig({ config: data.config }).catch((error: Error) => ({
+          const error: E.ErrorMessage | void = await tutorialConfig({ data }).catch((error: Error) => ({
             type: 'UnknownError',
             message: `Location: tutorial config.\n\n${error.message}`,
           }))
@@ -231,9 +231,8 @@ class Channel implements Channel {
           if (!tutorialContinue) {
             throw new Error('Invalid tutorial to continue')
           }
-          const continueConfig: TT.TutorialConfig = tutorialContinue.config
           await tutorialConfig({
-            config: continueConfig,
+            data: tutorialContinue,
             alreadyConfigured: true,
           })
           // update the current stepId on startup
@@ -307,7 +306,7 @@ class Channel implements Channel {
         await vscode.commands.executeCommand(COMMANDS.SET_CURRENT_POSITION, action.payload.position)
         await solutionActions({ actions: action.payload.actions, send: this.send })
         // run test following solution to update position
-        vscode.commands.executeCommand(COMMANDS.RUN_TEST)
+        vscode.commands.executeCommand(COMMANDS.RUN_TEST, { subtasks: true })
         return
       case 'EDITOR_SYNC_PROGRESS':
         // update progress when a level is deemed complete in the client
@@ -318,7 +317,7 @@ class Channel implements Channel {
         await showOutput(channel)
         return
       case 'EDITOR_RUN_TEST':
-        vscode.commands.executeCommand(COMMANDS.RUN_TEST)
+        vscode.commands.executeCommand(COMMANDS.RUN_TEST, action?.payload)
         return
       default:
         logger(`No match for action type: ${actionType}`)