+++ web-app/src/containers/Tutorial/components/Step.tsx
@@ -1,13 +1,14 @@
 import * as React from 'react'
 import * as T from 'typings'
 import { css, jsx } from '@emotion/core'
-import Icon from '../../../components/Icon'
+import TestStatusIcon from './TestStatusIcon'
 import Markdown from '../../../components/Markdown'
 
 interface Props {
   order: number
   content: string
   status: T.ProgressStatus
+  subtasks: { name: string; pass: boolean }[] | null
   onLoadSolution(): void
 }
 
@@ -23,6 +24,13 @@ const styles = {
   statusContainer: {
     paddingTop: 0,
     width: '1rem',
+  },
+  subtasks: {
+    marginTop: '1rem',
+  },
+  subtask: {
+    marginBottom: '1rem',
+    display: 'flex',
   },
   options: {
     display: 'flex' as 'flex',
@@ -42,11 +50,24 @@ const Step = (props: Props) => {
     <div>
       <div css={styles.card}>
         <div css={styles.statusContainer}>
-          {props.status === 'ACTIVE' && <Icon type="success-filling" size="small" style={{ color: 'lightgrey' }} />}
-          {props.status === 'COMPLETE' && <Icon type="success-filling" size="small" style={{ color: '#37B809' }} />}
+          {props.status === 'ACTIVE' && <TestStatusIcon size="small" />}
+          {props.status === 'COMPLETE' && <TestStatusIcon size="small" checked />}
         </div>
         <div>
-          <Markdown>{props.content || ''}</Markdown>
+          <div css={styles.content}>
+            <Markdown>{props.content || ''}</Markdown>
+          </div>
+          {props.subtasks ? (
+            <ul css={styles.subtasks}>
+              {props.subtasks.map((subtask) => (
+                <li key={subtask.name} css={styles.subtask}>
+                  <TestStatusIcon size="xs" checked={subtask.pass} />
+
+                  <span style={{ marginLeft: '0.5rem' }}>{subtask.name}</span>
+                </li>
+              ))}
+            </ul>
+          ) : null}
         </div>
       </div>
     </div>