+++ web-app/src/services/state/actions/testNotify.ts
@@ -22,11 +22,21 @@ const testActions: ActionFunctionMap<CR.MachineContext, CR.MachineEvent> = {
       type: 'warning',
       title: event.payload.fail.title,
       content: event.payload.fail.description,
+      summary: event.payload.fail.summary,
     }),
   }),
   // @ts-ignore
   testClear: assign({
     testStatus: null,
+  }),
+  // @ts-ignore
+  testSubtasks: assign({
+    testStatus: (context, event) => ({
+      type: 'hidden',
+      title: '',
+      content: '',
+      summary: event.payload.summary,
+    }),
   }),
 }