+++ web-app/stories/Level.stories.tsx
@@ -93,10 +93,94 @@ storiesOf('Level', module)
         index={0}
         status={level.status}
         steps={level.steps}
+        onRunTest={action('onRunTest')}
+        onOpenLogs={action('onOpenLogs')}
         onContinue={action('onContinue')}
         onLoadSolution={action('onLoadSolution')}
         processes={[]}
         testStatus={null}
+      />
+    )
+  })
+  .add('Subtasks', () => {
+    const level = {
+      id: 'L1',
+      index: 0,
+      title: 'A Title',
+      description: 'A summary of the level',
+      content: 'Some content here in markdown',
+      setup: null,
+      status: 'ACTIVE' as 'ACTIVE',
+      steps: [
+        {
+          id: 'L1:S1',
+          title: 'First Step',
+          content: 'First step description',
+          setup: {
+            id: 'L1:S1:SETUP',
+            commits: ['abcdefg'],
+          },
+          solution: {
+            id: 'L1:S1:SOLUTION',
+            commits: ['hijklmn'],
+          },
+          status: 'COMPLETE',
+        },
+        {
+          id: 'L1:S2',
+          title: 'Second Step',
+          content: 'Second step description',
+          setup: {
+            id: 'L1:S2:SETUP',
+            commits: ['abcdefg'],
+            subtasks: true,
+            filter: '^SomeTest',
+          },
+          solution: {
+            id: 'L1:S2:SOLUTION',
+            commits: ['hijklmn'],
+          },
+          status: 'ACTIVE',
+        },
+        {
+          id: 'L1:S3',
+          title: 'Third Step',
+          content: 'Third step description',
+          setup: {
+            id: 'L1:S3:SETUP',
+            commits: ['abcdefg'],
+          },
+          solution: {
+            id: 'L1:S3:SOLUTION',
+            commits: ['hijklmn'],
+          },
+          status: 'INCOMPLETE',
+        },
+      ],
+    }
+    const testStatus: T.TestStatus = {
+      type: 'error',
+      title: 'Test Failed because X',
+      summary: {
+        'The first task in a set of multiple subtasks': false,
+        'The second task out of a bunch of subtasks': true,
+        'The third and final task that has more text and might even wrap around because the text just keeps rambling on longer than anyone would conceivably want to read': false,
+      },
+    }
+    return (
+      <Level
+        menu={menu}
+        title={level.title}
+        content={level.content}
+        index={0}
+        status={level.status}
+        steps={level.steps}
+        onRunTest={action('onRunTest')}
+        onOpenLogs={action('onOpenLogs')}
+        onContinue={action('onContinue')}
+        onLoadSolution={action('onLoadSolution')}
+        processes={[]}
+        testStatus={testStatus}
       />
     )
   })
@@ -143,7 +227,6 @@ storiesOf('Level', module)
           status: 'COMPLETE',
         },
       ],
-      status: 'COMPLETE',
     }
     return (
       <Level
@@ -151,8 +234,10 @@ storiesOf('Level', module)
         title={level.title}
         content={level.content}
         index={0}
-        status={level.status}
+        status={'COMPLETE'}
         steps={level.steps}
+        onRunTest={action('onRunTest')}
+        onOpenLogs={action('onOpenLogs')}
         onContinue={action('onContinue')}
         onLoadSolution={action('onLoadSolution')}
         processes={[
@@ -242,6 +327,8 @@ storiesOf('Level', module)
         index={0}
         status={level.status}
         steps={level.steps}
+        onRunTest={action('onRunTest')}
+        onOpenLogs={action('onOpenLogs')}
         onContinue={action('onContinue')}
         onLoadSolution={action('onLoadSolution')}
         processes={[]}
@@ -320,6 +407,8 @@ storiesOf('Level', module)
         index={0}
         status={level.status}
         steps={level.steps}
+        onRunTest={action('onRunTest')}
+        onOpenLogs={action('onOpenLogs')}
         onContinue={action('onContinue')}
         onLoadSolution={action('onLoadSolution')}
         processes={[]}
@@ -346,6 +435,8 @@ storiesOf('Level', module)
         index={0}
         status={level.status}
         steps={level.steps}
+        onRunTest={action('onRunTest')}
+        onOpenLogs={action('onOpenLogs')}
         onContinue={action('onContinue')}
         onLoadSolution={action('onLoadSolution')}
         processes={[]}